Started to rebuild balances page
Removed plenty of switch statements, now we use a text file to get text elements in the correct language.
Improved the budget overview, now we can see not only the name of the budgets but also their balances.
Fixed a bug on the index page (recent spendings got appended multiple times).
Started to clean up balances page related files, but this is not done yet.